fix: resolve all QuickJS integration compile errors
- Remove duplicate JSResult/JSNativeCallback from breeze_api.h, include js_engine.h instead
- Use void* opaque pointer in JSEngine header, standalone JSEngineImpl in cpp
- Convert ConvertResult to static free function (no header dependency on QuickJS types)
- Use JSCFunctionMagic with unique magic numbers for native callback dispatch
- Replace JS_DeletePropertyStr (removed in latest QuickJS) with JS_SetPropertyStr
- Cast JSType to int32_t in C ABI for breeze_js_result

+++ src/breeze_api.cpp
@@ -265,13 +265,13 @@ breeze_js_result breeze_js_eval(const char* source, const char* filename) {
         source, filename ? filename : "<eval>");
     // NOTE: the returned value pointer is only valid until the next JS call
     // Callers should copy the string immediately.
-    return {r.success ? 1 : 0, r.type, r.value.c_str()};
+    return {r.success ? 1 : 0, static_cast<int32_t>(r.type), r.value.c_str()};
 }
 
 breeze_js_result breeze_js_eval_module(const char* source, const char* filename) {
     auto r = breeze::BreezeAPI::Instance().EvalJSModule(
         source, filename ? filename : "<module>.mjs");
-    return {r.success ? 1 : 0, r.type, r.value.c_str()};
+    return {r.success ? 1 : 0, static_cast<int32_t>(r.type), r.value.c_str()};
 }
 
 bool breeze_js_set_global_string(const char* name, const char* value) {
